graphql_users/schema.py
import graphene

from api_endpoint.schema import ApiQuery, ApiMutation


class Query(ApiQuery, graphene.ObjectType):
    # This class will inherit from multiple Queries
    # as we begin to add more apps to our project
    pass

class Mutation(ApiMutation, graphene.ObjectType):
    pass


schema = graphene.Schema(query=Query, mutation=Mutation)

