requirements.txt
django
Faker
psycopg2-binary
graphene-django


django-filter